+++ hotmart_python/hotmart.py
@@ -4,7 +4,9 @@ import logging
 import time
 import sys
 
-from typing import List, Dict, Any, Optional
+from typing import List, Dict, Any, Optional, Tuple
+
+Response = List[Dict[str, Any]]
 
 # Base Logging Configs
 logger = logging.getLogger(__name__)
@@ -34,33 +36,6 @@ logger.addHandler(hdlr=ch)
 logger.setLevel(level=logging.CRITICAL)
 
 
-class HTTPRequestException(Exception):
-    def __init__(self, message, status_code, url, response_body=None):
-        super().__init__(message)
-        self.status_code = status_code
-        self.url = url
-        self.response_body = response_body
-
-    def __str__(self):
-        return (f"HTTPRequestException: {self.args[0]}, "
-                f"Status Code: {self.status_code}, "
-                f"URL: {self.url}, Response Body: {self.response_body}")
-
-
-class RequestException(Exception):
-    def __init__(self, message, url):
-        super().__init__(message)
-        self.url = url
-
-    def __str__(self):
-        return f"RequestException: {self.args[0]}, URL: {self.url}"
-
-
-# URL Configs
-PRODUCTION_BASE_URL = "https://developers.hotmart.com/payments/api/"
-SANDBOX_BASE_URL = "https://sandbox.hotmart.com/payments/api/"
-
-
 class Hotmart:
     def __init__(self, client_id: str, client_secret: str, basic: str,
                  api_version: int = 1,
@@ -81,8 +56,7 @@ class Hotmart:
         self.secret = client_secret
         self.basic = basic
         self.sandbox = sandbox
-        self.base_url = SANDBOX_BASE_URL if sandbox else PRODUCTION_BASE_URL
-        self.base_url = f'{self.base_url}v{api_version}'
+        self.api_version = api_version
 
         # Token caching and some logic to do better logging.
         self.token_cache = None
@@ -91,12 +65,6 @@ class Hotmart:
 
         self.logger = logging.getLogger(__name__)
         self.logger.setLevel(log_level)
-
-    def _sandbox_error_warning(self):
-        self.logger.warning("At the date of last update for this library"
-                            " the Hotmart Sandbox API does NOT supported this method.")
-        self.logger.warning("This method probably won't work in the Sandbox mode.")
-        return
 
     @staticmethod
     def _build_payload(**kwargs: Any) -> Dict[str, Any]:
@@ -113,6 +81,61 @@ class Hotmart:
                 payload[key] = value
         return payload
 
+    @staticmethod
+    def _handle_response(response: Tuple[Dict[str, Any], List[Dict[str, Any]]],
+                         enhance=True) -> Response:
+        """
+        Standardizes the output of the response to always be a list of dictionaries.
+        :param response: The original response which can be a dictionary or a list of dictionaries.
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
+        :return: A list of dictionaries.
+        """
+        try:
+            if enhance:
+                if isinstance(response, dict) and response["items"]:
+                    return response["items"]
+
+                if isinstance(response, dict) and response["lessons"]:
+                    return response["lessons"]
+
+                if isinstance(response, list):
+                    return response
+
+            if not enhance:
+                if isinstance(response, dict):
+                    return [response]
+
+                if isinstance(response, list):
+                    return response
+
+            raise ValueError("Response must be a dictionary or a list of dictionaries.")
+        except KeyError:
+            return [response]
+
+    def _build_url(self, endpoint_type: str):
+        """
+        Builds the URLs for better dynamic requests.
+        :param endpoint_type: Supported types: payments or club
+        :return:
+        """
+
+        supported_types = ['payments', 'club']
+        if endpoint_type.lower() not in supported_types:
+            raise ValueError(f"Unsupported endpoint type: {endpoint_type}")
+
+        return (f'https://{"sandbox" if self.sandbox else "developers"}.hotmart.com/'
+                f'{endpoint_type.lower()}/api/v{self.api_version}')
+
+    def _sandbox_error_warning(self):
+        """
+        Logs a warning message about the Hotmart Sandbox API not supporting some requests
+        :return:
+        """
+        self.logger.warning("At the date of last update for this library"
+                            " the Hotmart Sandbox API does NOT supported this method.")
+        self.logger.warning("This method probably won't work in the Sandbox mode.")
+        return
+
     def _log_instance_mode(self) -> None:
         """
         Logs the instance mode (Sandbox or Production).
@@ -121,11 +144,13 @@ class Hotmart:
         return self.logger.warning(
             f"Instance in {'Sandbox' if self.sandbox else 'Production'} mode")
 
-    def _make_request(self, method: Any, url: str,
+    def _make_request(self,
+                      method: Any,
+                      url: str,
                       headers: Optional[Dict[str, str]] = None,
                       params: Optional[Dict[str, Any]] = None,
-                      body: Optional[Dict[str, str]] = None,
-                      log_level: int = None) -> Optional[Dict[str, Any]]:
+                      body: Optional[Dict[str, str]] = None, log_level: Optional[int] = None) -> \
+            List[Dict[str, Any]]:
         """
         Makes a request to the given url.
         :param method: The request method (e.g, requests.get, requests.post).
@@ -141,54 +166,49 @@ class Hotmart:
         if log_level is not None:
             logger = logging.getLogger(__name__)  # noqa
             logger.setLevel(log_level)
-        try:
-            self.logger.debug(f"Request URL: {url}")
-            self.logger.debug(f"Request headers: {headers}")
-            self.logger.debug(f"Request params: {params}")
-            self.logger.debug(f"Request body: {body}")
 
+        self.logger.debug(f"Request URL: {url}")
+        self.logger.debug(f"Request headers: {headers}")
+        self.logger.debug(f"Request params: {params}")
+        self.logger.debug(f"Request body: {body}")
+
+        try:
             response = method(url, headers=headers, params=params, data=body)
             self.logger.debug(f"Response content: {response.text}")
+            if response.status_code == requests.codes.ok:
+                return response.json()
             response.raise_for_status()
-            return response.json()
 
-        except requests.exceptions.HTTPError:
+        except requests.exceptions.HTTPError as e:
             # noinspection PyUnboundLocalVariable
-            if response.status_code != 403:
-                raise HTTPRequestException("HTTP Error",
-                                           response.status_code,
-                                           url,
-                                           body)
-            error_message = "Forbidden."
-            self.logger.error(f"Error {response.status_code}: {error_message}")
-            if self.sandbox:
-                self.logger.error(
-                    "Check if the provided credentials are for Sandbox mode.")
+            if e.response.status_code == 401 or e.response.status_code == 403:
+                if self.sandbox:
+                    self.logger.error("Perhaps the credentials aren't for Sandbox Mode?")
+                else:
+                    self.logger.error("Perhaps the credentials are for Sandbox Mode?")
+                raise e
 
-            self.logger.error(
-                "Perhaps the provided credentials are for Sandbox mode?")
-            raise HTTPRequestException(error_message, response.status_code,
-                                       url, body)
+            if e.response.status_code == 422:
+                self.logger.error(f"Error {e.response.status_code}")
+                self.logger.error("This usually happens when the request is missing"
+                                  " body parameters.")
+                raise e
 
-        except requests.exceptions.RequestException as e:
-            error_message = str(e) if str(
-                e) else "An error occurred while making the request"
-            self.logger.error(
-                f"Error making request to {url}: {error_message}")
-            raise RequestException(
-                f"Error making request to {url}: {error_message}",
-                url)
+            if e.response.status_code == 500 and self.sandbox:
+                self.logger.error("This happens with some endpoints in the Sandbox Mode.")
+                self.logger.error("Usually the API it's not down, it's just a bug.")
+
+            raise e
 
     def _is_token_expired(self) -> bool:
         """
         Checks if the current token has expired.
         :return: True if the token has expired, False otherwise.
         """
-        return (self.token_expires_at is not None and self.token_expires_at <
-                time.time())
+        return self.token_expires_at is not None and self.token_expires_at < time.time()
 
     def _fetch_new_token(self) -> str:
-        self.logger.info("Fetching a new access token.")
+        self.logger.debug("Fetching a new access token.")
 
         method_url = 'https://api-sec-vlc.hotmart.com/security/oauth/token'
         headers = {'Authorization': self.basic}
@@ -197,21 +217,21 @@ class Hotmart:
 
         response = self._make_request(requests.post, method_url,
                                       headers=headers, params=payload)
-        self.logger.info("Token obtained successfully")
+        self.logger.debug("Token obtained successfully")
         return response['access_token']
 
-    def _get_token(self) -> Optional[str]:
+    def _get_token(self) -> str:
         """
         Retrieves an access token to authenticate requests.
         :return: The access token if obtained successfully, otherwise None.
         """
         if not self._is_token_expired() and self.token_cache is not None:
             if not self.token_found_in_cache:
-                self.logger.info("Token found in cache.")
+                self.logger.debug("Token found in cache.")
                 self.token_found_in_cache = True
             return self.token_cache
 
-        self.logger.info("Token not found in cache or expired.")
+        self.logger.debug("Token not found in cache or expired.")
 
         token = self._fetch_new_token()
         if token is not None:
@@ -219,8 +239,9 @@ class Hotmart:
             self.token_found_in_cache = False
         return token
 
-    def _request_with_token(self, method: str, url: str, body: Optional[Dict[str, Any]] = None,
-                            params: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
+    def _request_with_token(self, method: str, url: str, enhance: bool = None,
+                            body: Optional[Dict[str, Any]] = None,
+                            params: Optional[Dict[str, Any]] = None) -> Response:
         """
         Makes an authenticated request (GET, POST, PATCH, etc.) to the
         specified URL with the given body or params.
@@ -248,181 +269,139 @@ class Hotmart:
         if method.upper() not in method_mapping:
             raise ValueError(f"Unsupported method: {method}")
 
-        return self._make_request(method_mapping[method.upper()], url, headers=headers,
-                                  params=params, body=body)
+        response = self._make_request(method_mapping[method.upper()], url, headers=headers,
+                                      params=params, body=body)
 
-    def _pagination(self, method: str, url: str, params: Optional[Dict[str, Any]] = None,
-                    paginate: bool = False) -> Optional[List[Dict[str, Any]]]:
-        """
-        Retrieves all pages of data for a paginated endpoint.
-        :param url: The URL of the paginated endpoint.
-        :param params: Optional request parameters.
-        :param paginate: Whether to paginate the results or not
-         (default is False).
-        :return: A list containing data from all pages,
-         or None if an error occurred.
-        """
-        if not paginate:
-            response = self._request_with_token(method=method, url=url, params=params)
-            return response.get("items", []) if response else None
-        all_items = []
+        return self._handle_response(response, enhance=enhance)
 
-        self.logger.info("Fetching first page...")
-        response = self._request_with_token(method=method, url=url, params=params)
-
-        if response is None:
-            raise ValueError("Failed to fetch first page.")
-
-        all_items.extend(response.get("items", []))
-
-        while "next_page_token" in response.get("page_info", {}):
-            next_page_token = response["page_info"]["next_page_token"]
-            self.logger.info(f"Fetching next page with token: {next_page_token}")
-            params["page_token"] = next_page_token
-            response = self._request_with_token(method, url, params=params)
-
-            if response is None:
-                raise ValueError(
-                    f"Failed to fetch next page with token: {next_page_token}")
-
-            all_items.extend(response.get("items", []))
-
-        self.logger.info("Finished fetching all pages.")
-        return all_items
-
-    def get_sales_history(self, paginate: bool = False, **kwargs: Any) -> Optional[Dict[str, Any]]:
+    def get_sales_history(self, enhance: bool = True, **kwargs: Any) -> Response:
         """
         Retrieves sales history data based on the provided filters.
-
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param kwargs: Filters to apply on the request. Expected kwargs can be found in
         the "Request parameters" section of the API Docs.
-        :param paginate: Whether to paginate the results or not (default is False).
         :return: Sales history data if available, otherwise None.
         """
+
         self._log_instance_mode()
 
         method = "get"
-        url = f'{self.base_url}/sales/history'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/sales/history'
         payload = self._build_payload(**kwargs)
-        return self._pagination(method=method, url=url, params=payload,
-                                paginate=paginate)
+        return self._request_with_token(method=method, url=url, params=payload, enhance=enhance)
 
-    def get_sales_summary(self, paginate: bool = False, **kwargs: Any) -> Optional[Dict[str, Any]]:
+    def get_sales_summary(self, enhance: bool = True, **kwargs: Any) -> Response:
         """
         Retrieves sales summary data based on the provided filters.
-
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param kwargs: Filters to apply on the request. Expected kwargs can be found in the
         "Request parameters" section of the API Docs.
-        :param paginate: Whether to paginate the results or not (default is False).
         :return: Sales summary data if available, otherwise None.
         """
 
         self._log_instance_mode()
 
         method = "get"
-        url = f'{self.base_url}/sales/summary'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/sales/summary'
         payload = self._build_payload(**kwargs)
-        return self._pagination(method=method, url=url, params=payload,
-                                paginate=paginate)
+        return self._request_with_token(method=method, url=url, params=payload, enhance=enhance)
 
-    def get_sales_participants(self, paginate: bool = True, **kwargs: Any) -> \
-            Optional[Dict[str, Any]]:
+    def get_sales_participants(self, enhance: bool = True, **kwargs: Any) -> Response:
         """
         Retrieves sales user data based on the provided filters.
-
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param kwargs: Filters to apply on the request. Expected kwargs can be found in the
         "Request parameters" section of the API Docs.
-        :param paginate: Whether to paginate the results or not (default is True).
         :return: Sales user data if available, otherwise None.
         """
 
         self._log_instance_mode()
 
         method = "get"
-        url = f'{self.base_url}/sales/users'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/sales/users'
         payload = self._build_payload(**kwargs)
-        return self._pagination(method=method, url=url, params=payload, paginate=paginate)
+        return self._request_with_token(method=method, url=url, params=payload, enhance=enhance)
 
-    def get_sales_commissions(self, paginate: bool = False, **kwargs: Any) -> \
-            Optional[Dict[str, Any]]:
+    def get_sales_commissions(self, enhance: bool = True, **kwargs: Any) -> Response:
         """
         Retrieves sales commissions data based on the provided filters.
-
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param kwargs: Filters to apply on the request. Expected kwargs can be found in the
         "Request parameters" section of the API Docs.
-        :param paginate: Whether to paginate the results or not (default is False).
         :return: Sales commissions data if available, otherwise None.
         """
 
         self._log_instance_mode()
 
         method = "get"
-        url = f'{self.base_url}/sales/commissions'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/sales/commissions'
         payload = self._build_payload(**kwargs)
-        return self._pagination(method=method, url=url, params=payload, paginate=paginate)
+        return self._request_with_token(method=method, url=url, params=payload, enhance=enhance)
 
-    def get_sales_price_details(self, paginate: bool = False, **kwargs: Any) \
-            -> Optional[Dict[str, Any]]:
+    def get_sales_price_details(self, enhance: bool = True, **kwargs: Any) -> Response:
         """
         Retrieves sales price details based on the provided filters.
-
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param kwargs: Filters to apply on the request. Expected kwargs can be found in the
         "Request parameters" section of the API Docs.
-        :param paginate: Whether to paginate the results or not (default is False).
         :return: Sales price details if available, otherwise None.
         """
 
         self._log_instance_mode()
 
         method = "get"
-        url = f'{self.base_url}/sales/price/details'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/sales/price/details'
         payload = self._build_payload(**kwargs)
-        return self._pagination(method=method, url=url, params=payload, paginate=paginate)
+        return self._request_with_token(method=method, url=url, params=payload, enhance=enhance)
 
-    def get_subscriptions(self, paginate: bool = False, **kwargs: Any) -> \
-            Optional[Dict[str, Any]]:
+    def get_subscriptions(self, enhance: bool = True, **kwargs: Any) -> Response:
         """
         Retrieves subscription data based on the provided filters.
-
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param kwargs: Filters to apply on the request. Expected kwargs can be found in the
         "Request parameters" section of the API Docs.
-        :param paginate: Whether to paginate the results or not (default is False).
         :return: Subscription data if available, otherwise None.
         """
 
         self._log_instance_mode()
 
         method = "get"
-        url = f'{self.base_url}/subscriptions'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/subscriptions'
         payload = self._build_payload(**kwargs)
-        return self._pagination(method=method, url=url, params=payload, paginate=paginate)
+        return self._request_with_token(method=method, url=url, params=payload, enhance=enhance)
 
-    def get_subscriptions_summary(self, paginate: bool = False, **kwargs: Any) -> \
-            Optional[Dict[str, Any]]:
+    def get_subscriptions_summary(self, enhance: bool = True, **kwargs: Any) -> Response:
         """
         Retrieves subscription summary data based on the provided filters.
-        :param paginate: Whether to paginate the results or not (default is False).
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param kwargs: Filters to apply on the request. Expected kwargs can be found in the
         "Request parameters" section of the API Docs.
         :return: Subscription data if available, otherwise None.
         """
 
         self._log_instance_mode()
+        self._sandbox_error_warning()
 
         method = "get"
-        url = f'{self.base_url}/subscriptions/summary'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/subscriptions/summary'
         payload = self._build_payload(**kwargs)
-        return self._pagination(method=method, url=url, params=payload, paginate=paginate)
+        return self._request_with_token(method=method, url=url, params=payload, enhance=enhance)
 
-    def get_subscription_purchases(self, subscriber_code, paginate: bool = False, **kwargs: Any) ->\
-            Optional[Dict[str, Any]]:
+    def get_subscription_purchases(self, subscriber_code, enhance: bool = True, **kwargs: Any) -> \
+            Response:
         """
         Retrieves subscription purchases data based on the provided filters.
-
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param subscriber_code: The subscriber code to filter the request.
         :param kwargs: Filters to apply on the request. Expected kwargs can be found in the
         "Request parameters" section of the API Docs.
-        :param paginate: Whether to paginate the results or not (default is False).
         :return: Subscription purchases data if available, otherwise None.
         """
 
@@ -431,12 +410,12 @@ class Hotmart:
             self._sandbox_error_warning()
 
         method = "get"
-        url = f'{self.base_url}/subscriptions/{subscriber_code}/purchases'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/subscriptions/{subscriber_code}/purchases'
         payload = self._build_payload(**kwargs)
-        return self._pagination(method=method, url=url, params=payload, paginate=paginate)
+        return self._request_with_token(method=method, url=url, params=payload, enhance=enhance)
 
-    def cancel_subscription(self, subscriber_code: list[str], send_email: bool = True) ->\
-            Optional[Dict[str, Any]]:
+    def cancel_subscription(self, subscriber_code: list[str], send_email: bool = True) -> Response:
         """
         Cancels a subscription.
 
@@ -450,15 +429,16 @@ class Hotmart:
             self._sandbox_error_warning()
 
         method = "post"
-        url = f'{self.base_url}/subscriptions/cancel'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/subscriptions/cancel'
         payload = {
             "subscriber_code": subscriber_code,
             "send_email": send_email
         }
         return self._request_with_token(method=method, url=url, body=payload)
 
-    def reactivate_and_charge_subscription(self, subscriber_code: list[str], charge: bool = False)\
-            -> Optional[Dict[str, Any]]:
+    def reactivate_and_charge_subscription(self, subscriber_code: list[str], charge: bool = False) \
+            -> Response:
         """
         Reactivates and charges a subscription.
 
@@ -473,14 +453,15 @@ class Hotmart:
             self._sandbox_error_warning()
 
         method = "post"
-        url = f'{self.base_url}/subscriptions/reactivate'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/subscriptions/reactivate'
         payload = {
             "subscriber_code": subscriber_code,
             "charge": charge
         }
         return self._request_with_token(method=method, url=url, body=payload)
 
-    def change_due_day(self, subscriber_code: str, new_due_day: int) -> Optional[Dict[str, Any]]:
+    def change_due_day(self, subscriber_code: str, new_due_day: int) -> Response:
         """
         Changes the due day of a subscription.
 
@@ -494,18 +475,16 @@ class Hotmart:
             self._sandbox_error_warning()
 
         method = "patch"
-        url = f'{self.base_url}/subscriptions/change-due-day'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/subscriptions/{subscriber_code}'
         payload = {
-            "subscriber_code": subscriber_code,
-            "new_due_day": new_due_day
+            "due_day": new_due_day
         }
         return self._request_with_token(method=method, url=url, body=payload)
 
-    def create_coupon(self, product_id: str, coupon_code: str, discount: float) -> \
-            Optional[Dict]:
+    def create_coupon(self, product_id: str, coupon_code: str, discount: float) -> Response:
         """
         Creates a coupon for a product.
-
         :param product_id: UID of the product you want to create the coupon for.
         :param coupon_code: The code of the coupon you want to create.
         :param discount: The discount you want to apply to the coupon, must be greater than 0 and
@@ -518,18 +497,20 @@ class Hotmart:
             self._sandbox_error_warning()
 
         method = "post"
-        url = f'{self.base_url}/product/{product_id}/coupon'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/product/{product_id}/coupon'
         payload = {
             "code": coupon_code,
             "discount": discount
         }
         return self._request_with_token(method=method, url=url, body=payload)
 
-    def get_coupon(self, product_id: str) -> Optional[Dict[str, Any]]:
+    def get_coupon(self, product_id: str, code: str, enhance: bool = True) -> Response:
         """
         Retrieves a coupon for a product.
-
+        :param enhance: When True, discards page_info and returns only the items. (Default is True)
         :param product_id: UID of the product you want to retrieve the coupon for.
+        :param code: The code of the coupon you want to retrieve.
         :return: All Coupons for the product.
         """
 
@@ -538,8 +519,12 @@ class Hotmart:
             self._sandbox_error_warning()
 
         method = "get"
-        url = f'{self.base_url}/coupon/product/{product_id}'
-        return self._request_with_token(method=method, url=url)
+        base_url = self._build_url('payments')
+        url = f'{base_url}/coupon/product/{product_id}'
+        params = {
+            "code": code
+        }
+        return self._request_with_token(method=method, url=url, params=params, enhance=enhance)
 
     def delete_coupon(self, coupon_id):
         """
@@ -553,5 +538,6 @@ class Hotmart:
             self._sandbox_error_warning()
 
         method = "delete"
-        url = f'{self.base_url}/coupon/{coupon_id}'
+        base_url = self._build_url('payments')
+        url = f'{base_url}/coupon/{coupon_id}'
         return self._request_with_token(method=method, url=url)